hil(GOAL 8 C/D/E): 上位脫困、近距離安全區、上位與模擬同一個時鐘
E 線(時鐘):driver 用 odom 框包的 t_ms 發 /clock、ROS 端 use_sim_time,時戳與 TF 都改吃
模擬時間;SIM_TIME=0 整組回牆鐘當對照(只切節點而時戳仍是模擬時間是兩邊都不對的組合,
拿它當「改前」等於灌水)。閒時兩邊各 3/3 綠——問題不重現;ROS 容器壓 1 核也不重現;
Renode 壓 1 核(模擬 0.41× → 0.2×)才重現:模擬時鐘 5/5 綠、牆鐘 4 綠 1 紅(紅的那次完全
沒解鎖、離 goal 3598 mm)。踩到一個互等:Nav2 在 use_sim_time 下要等有效時鐘才啟動,而
時鐘要模擬跑起來才有 → driver 改成一啟動就用牆鐘 timer 發 /clock(單趟從卡 4 小時變 3 分鐘)。

D 線(近距離安全區):測距走 CAN 0x301(u16 mm),韌體 SAFETY_ZONE 在 700 mm 線性限速、
350 mm 不准前進(只擋前進、允許後退,同保險桿);C18 用專屬的 --scene zone——接觸場景
head-on/long-slip 預設關掉這層,否則 C13/C16 要的碰撞不會發生。C18 綠(間隙 149 mm、
碰撞 0),zone-off 紅(間隙 0、碰撞 249 步)。blind-scan 下雷射與這顆一起蒙,C11 仍紅。

C 線(上位脫困):bump_layer 獨立圖層(蒙住的雷射回報「空曠」,標在同一層會被它的清除
射線抹掉)、後退用里程計閉環(照牆鐘數秒在 0.35× 模擬下只退到三分之一)、標記放大成
車寬 + 餘裕。300 s 下脫困 4 次、Nav2 succeeded、真值到 goal 72 mm(RECOVER=0 卡在原地
離 goal 2279 mm);C17 仍紅——最後一次 SLIP 之後還有 3 步輕觸,判準不為了變綠放寬。

過程中修掉三個既有缺陷:
- driver 把安全旗標發成 UInt8,而 SLIP 是 bit 8 → 上位永遠看不到打滑(C12 用的 bit 7
  剛好在界線內,所以拖到現在才浮出來)。改 UInt16。
- 假受控體的碰撞判定用「更新前」的車速試算位置:停在障礙物上時 v_body=0 → 試算位置
  就是現在的位置 → 永遠碰撞 → 連倒車都退不開。症狀是 odom 報退了 1.25 m、真值 20 s
  沒動。改成用這一步的力算出的新車速試算(Rust 與 Python 受控體同步)。
- C13 的子判準「SLIP 亮 0.5 s 後輪子停」前提被 C 線推翻(脫困本來就會再動)→ 改成
  「0.5 s 內輪子停過一次」,門檻不動、no-latch 負對照仍紅。

+++ examples/hil-stm32/ros/scan_check.py
@@ -6,7 +6,7 @@ import sys
 import rclpy
 from rclpy.node import Node
 from sensor_msgs.msg import LaserScan
-from std_msgs.msg import UInt8
+from std_msgs.msg import UInt16
 
 
 class ScanCheck(Node):
@@ -17,7 +17,7 @@ class ScanCheck(Node):
         self.last = None
         self.flags = set()
         self.create_subscription(LaserScan, "scan", self.on_scan, 10)
-        self.create_subscription(UInt8, "hil/safety_flags", lambda m: self.flags.add(m.data), 10)
+        self.create_subscription(UInt16, "hil/safety_flags", lambda m: self.flags.add(m.data), 10)
         self.create_timer(float(self.get_parameter("seconds").value), self.done)
 
     def on_scan(self, m):
